+++ src/main/java/me/zhyd/oauth/config/AuthDefaultSource.java
@@ -73,7 +73,7 @@ public enum AuthDefaultSource implements AuthSource {
         }
     },
     /**
-     * 钉钉
+     * 钉钉扫码登录
      */
     DINGTALK {
         @Override
@@ -89,6 +89,25 @@ public enum AuthDefaultSource implements AuthSource {
         @Override
         public String userInfo() {
             return "https://oapi.dingtalk.com/sns/getuserinfo_bycode";
+        }
+    },
+    /**
+     * 钉钉账号登录
+     */
+    DINGTALK_ACCOUNT {
+        @Override
+        public String authorize() {
+            return "https://oapi.dingtalk.com/connect/oauth2/sns_authorize";
+        }
+
+        @Override
+        public String accessToken() {
+            return DINGTALK.accessToken();
+        }
+
+        @Override
+        public String userInfo() {
+            return DINGTALK.userInfo();
         }
     },
     /**
